Move action out of svelte
